+++ conversorDeMoeda/conversorDeMoeda/Model/Moeda.swift
@@ -1,0 +1,15 @@
+//
+//  Moeda.swift
+//  conversorDeMoeda
+//
+//  Created by Diogenes de Souza on 21/01/21.
+//
+
+import Foundation
+
+struct Moeda {
+    var nome:String?
+    var sigla:String?
+    var valor:Double?
+}
+
